scripts/gcs_to_snowflake/requirements.txt
pandas
google-cloud-storage
snowflake-connector-python[pandas]
python-dotenv